JCB_File_Validate

+++ app/Http/Controllers/JCBcontroller.php
@@ -17,38 +17,45 @@ class JCBcontroller extends Controller
         request()->validate([
             "jcb"=>"required|mimes:txt",
         ]);
-        $files=file(request('jcb')->getRealPath());
-        $data=(array_slice($files,0));
-        $parts =(array_chunk($data, 5000));
-        foreach($parts as $part){
-            $name=request('jcb');
-            $filename=$name->getClientoriginalName();
-            $fileName=uniqid()."_".$name->getClientoriginalName();
-            $fileSave=resource_path('file/jcb/'.'JCB'.$fileName.'.txt');
-            file_put_contents($fileSave,$part);
-        }
-        $path=$fileSave;
-        $filePath=glob($path);
-            foreach($filePath as $file){
-                $data=array_map('str_getcsv',file($file));
-                DB::delete('delete from jcbs');
-                    foreach($data as $row){
-                        jcb::Create([
-                            'Field1'=>$row [0],
-                            'filename'=>$filename,
-                        ]);
-                    };
-            }
-                $data=new jcb();
-                return view('jcbupload', [
-                    'users' => $data->show(),
-                    'one'   => $data->total1(),
-                    'two'   => $data->total2(),
-                    'three'   => $data->total3(),
-                    'four'   => $data->total4()
-                 ]);
+        $name=request('jcb');
+        $filename=$name->getClientoriginalName();
+        $e=substr($filename,0,7);
+        if  ($e==="JCB_CLR")        
+            {   $files=file(request('jcb')->getRealPath());
+                $data=(array_slice($files,0));
+                $parts =(array_chunk($data, 5000));
+                foreach($parts as $part){
+                    $name=request('jcb');
+                    $filename=$name->getClientoriginalName();
+                    $fileName=uniqid()."_".$name->getClientoriginalName();
+                    $fileSave=resource_path('file/jcb/'.'JCB'.$fileName.'.txt');
+                    file_put_contents($fileSave,$part);
+                }
+                $path=$fileSave;
+                $filePath=glob($path);
+                    foreach($filePath as $file){
+                        $data=array_map('str_getcsv',file($file));
+                        DB::delete('delete from jcbs');
+                            foreach($data as $row){
+                                jcb::Create([
+                                    'Field1'=>$row [0],
+                                    'filename'=>$filename,
+                                ]);
+                        };
+                }
+                    $data=new jcb();
+                    return view('jcbupload', [
+                        'users' => $data->show(),
+                        'one'   => $data->total1(),
+                        'two'   => $data->total2(),
+                        'three'   => $data->total3(),
+                        'four'   => $data->total4()
+                    ]);
+            }else
+                {
+                    return back()->with('error'," Doesn't work this file (Please select other file).");
+                }
     }
-
     public function download(){
         $data=new jcb();
         $users=$data->show();
